assistant/requirements.txt
# AI Assistant v1 dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
anthropic>=0.21.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
pydantic>=2.5.0
httpx>=0.27.0
supervisor>=4.2.0

# Development and testing
pytest-benchmark>=4.0.0
